added pause function, fixed authentication bug
pause function works using the fast forward button. I will be changing the pause function to skip.
