+++ QR/app/src/main/AndroidManifest.xml
@@ -15,10 +15,15 @@
         <activity
             android:name=".QueueActivity"
             android:label="@string/app_name" />
+
         <activity
             android:name=".SearchActivity"
             android:label="@string/app_name" />
-        <activity android:name=".MainActivity">
+        <activity
+        android:name=".MainActivity"
+        android:label="@string/app_name" />
+        <activity android:name=".SpotifyService">
+
             <intent-filter>
                 <action android:name="android.intent.action.VIEW"/>
