+++ QR/app/src/main/java/com/example/qr/MainActivity.java
@@ -40,22 +40,13 @@ public class MainActivity extends AppCompatActivity {
     SeekBar seekBar;
     BottomNavigationView btmNav;
 
-    private static final String CLIENT_ID = "c51c441a5bf749a4bf9a1a9b7987173a";
-    private static final String REDIRECT_URI = "http://qr.example.com";
-    protected SpotifyAppRemote mSpotifyAppRemote;
-    private static final int REQUEST_CODE = 1337;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        AuthenticationRequest.Builder builder =
-                new AuthenticationRequest.Builder(CLIENT_ID, AuthenticationResponse.Type.TOKEN, REDIRECT_URI);
-        builder.setScopes(new String[]{"app-remote-control","playlist-modify-public","user-read-currently-playing","user-modify-playback-state"});
-        AuthenticationRequest request = builder.build();
 
-        AuthenticationClient.openLoginActivity(this, REQUEST_CODE, request);
 
         queueName = findViewById(R.id.tvQueueName);
         songTitle = findViewById(R.id.tvQSongTitle);
@@ -72,29 +63,30 @@ public class MainActivity extends AppCompatActivity {
             public boolean onNavigationItemSelected(@NonNull MenuItem item) {
                 switch (item.getItemId()) {
                     case R.id.searchIcon:
-                        Intent intent = new Intent(MainActivity.this,SearchActivity.class);
+                        Intent intent = new Intent(MainActivity.this, SearchActivity.class);
                         startActivity(intent);
                         return true;
                     case R.id.homeIcon:
                         //stay home
                         return true;
                     case R.id.playIcon:
-                        Intent intenttwo = new Intent(MainActivity.this,QueueActivity.class);
+                        Intent intenttwo = new Intent(MainActivity.this, QueueActivity.class);
                         startActivity(intenttwo);
                         return true;
-                    default: return true;
+                    default:
+                        return true;
                 }
             }
         });
         songList = new ArrayList<>();
 
 
-        songList.add(new Song("Album Cover","Song title", "Artist", "Album"));
-        songList.add(new Song("Album Cover","Song title", "Artist", "Album"));
-        songList.add(new Song("Album Cover","Song title", "Artist", "Album"));
-        songList.add(new Song("Album Cover","Song title", "Artist", "Album"));
-        songList.add(new Song("Album Cover","Song title", "Artist", "Album"));
-        songList.add(new Song("Album Cover","Song title", "Artist", "Album"));
+        songList.add(new Song("Album Cover", "Song title", "Artist", "Album"));
+        songList.add(new Song("Album Cover", "Song title", "Artist", "Album"));
+        songList.add(new Song("Album Cover", "Song title", "Artist", "Album"));
+        songList.add(new Song("Album Cover", "Song title", "Artist", "Album"));
+        songList.add(new Song("Album Cover", "Song title", "Artist", "Album"));
+        songList.add(new Song("Album Cover", "Song title", "Artist", "Album"));
 
         SongAdapter songAdapter = new SongAdapter(this, (ArrayList<Song>) songList);
 
@@ -104,77 +96,10 @@ public class MainActivity extends AppCompatActivity {
         rvUsers.setAdapter(songAdapter);
         rvUsers.setLayoutManager(new LinearLayoutManager(this));
 
-
-
     }
-    protected void onActivityResult(int requestCode, int resultCode, Intent intent) {
-        super.onActivityResult(requestCode, resultCode, intent);
-
-        // Check if result comes from the correct activity
-        if (requestCode == REQUEST_CODE) {
-            AuthenticationResponse response = AuthenticationClient.getResponse(resultCode, intent);
-
-            switch (response.getType()) {
-                // Response was successful and contains auth token
-                case TOKEN:
-                    // Handle successful response
-                    break;
-
-                // Auth flow returned an error
-                case ERROR:
-                    // Handle error response
-                    break;
-
-                // Most likely auth flow was cancelled
-                default:
-                    // Handle other cases
-            }
-        }
-    }
-
     @Override
-    protected void onStart() {
-        super.onStart();
-
-
-        SpotifyAppRemote.disconnect(mSpotifyAppRemote);
-
-        ConnectionParams connectionParams =
-                new ConnectionParams.Builder(CLIENT_ID)
-                        .setRedirectUri(REDIRECT_URI)
-                        .showAuthView(true)
-                        .build();
-
-        SpotifyAppRemote.connect(this, connectionParams,
-                new Connector.ConnectionListener() {
-
-                    @Override
-                    public void onConnected(SpotifyAppRemote spotifyAppRemote) {
-                        mSpotifyAppRemote = spotifyAppRemote;
-                        Log.d("MainActivity", "Connected! Yay!");
-
-
-                        // Now you can start interacting with App Remote
-                        connected();
-                    }
-
-                    @Override
-                    public void onFailure(Throwable throwable) {
-                        Log.e("MainActivity", throwable.getMessage(), throwable);
-
-                        // Something went wrong when attempting to connect! Handle errors here
-                    }
-                });
-    }
-
-    private void connected() {
-
-    }
-
-    @Override
-    protected void onStop() {
+    protected void onStop(){
         super.onStop();
-        SpotifyAppRemote.disconnect(mSpotifyAppRemote);
     }
 }
