+++ QR/app/src/main/java/com/example/qr/SpotifyService.java
@@ -1,56 +1,133 @@
 package com.example.qr;
 
-public class SpotifyService {
- /*   private static final String CLIENT_ID = "c51c441a5bf749a4bf9a1a9b7987173a";
-    private static final String REDIRECT_URI = "http://qr.example.com";
-    private static final int REQUEST_CODE = 1337;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
+import android.widget.TextView;
+import android.widget.ImageView;
 
-    AuthenticationRequest.Builder builder =
-            new AuthenticationRequest.Builder(CLIENT_ID, AuthenticationResponse.Type.TOKEN, REDIRECT_URI);
-    builder.setScopes(new String[]{"app-remote-control","playlist-modify-public","user-read-currently-playing","user-modify-playback-state"});
-    AuthenticationRequest request = builder.build();
+import com.example.qr.Models.Song;
+import com.spotify.sdk.android.authentication.AuthenticationRequest;
+import com.spotify.sdk.android.authentication.AuthenticationResponse;
+import com.spotify.sdk.android.authentication.AuthenticationClient;
+import com.spotify.android.appremote.api.ConnectionParams;
+import com.spotify.android.appremote.api.Connector;
+import com.spotify.android.appremote.api.SpotifyAppRemote;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+import androidx.annotation.NonNull;
+import android.view.MenuItem;
 
-    AuthenticationClient.openLoginActivity(this, REQUEST_CODE, request);
-    }
+import android.content.Intent;
+
+import android.os.Bundle;
+import android.widget.Button;
+import android.view.View;
+
+import android.util.Log;
+import java.util.ArrayList;
+import java.util.List;
+
+public class SpotifyService extends MainActivity {
+
+    protected static SpotifyAppRemote mSpotifyAppRemote;
+    TextView tvMoment;
+    ImageView ivIcon;
 
     private static final String CLIENT_ID = "c51c441a5bf749a4bf9a1a9b7987173a";
     private static final String REDIRECT_URI = "http://qr.example.com";
-
-    SpotifyApi spotifyApi = new SpotifyApi.Builder()
-            .setAccessToken("taHZ2SdB-bPA3FsK3D7ZN5npZS47cMy-IEySVEGttOhXmqaVAIo0ESvTCLjLBifhHOHOIuhFUKPW1WMDP7w6dj3MAZdWT8CLI2MkZaXbYLTeoDvXesf2eeiLYPBGdx8tIwQJKgV8XdnzH_DONk")
-            .build();
-
     private static final int REQUEST_CODE = 1337;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_search);
+        setContentView(R.layout.activity_auth);
 
+        tvMoment = findViewById(R.id.tvMoment);
+        ivIcon = findViewById(R.id.queueIcon);
 
+        authenticateSpotify();
+
+        Intent intentActivity = new Intent(this,MainActivity.class);
+        startActivity(intentActivity);
     }
 
+    private void authenticateSpotify(){
+        AuthenticationRequest.Builder builder =
+                new AuthenticationRequest.Builder(CLIENT_ID, AuthenticationResponse.Type.TOKEN, REDIRECT_URI);
+        builder.setScopes(new String[]{"app-remote-control","playlist-modify-public","user-read-currently-playing","user-modify-playback-state"});
+        AuthenticationRequest request = builder.build();
+        AuthenticationClient.openLoginActivity(this, REQUEST_CODE, request);
+    }
+    protected void onActivityResult(int requestCode, int resultCode, Intent intent) {
+        super.onActivityResult(requestCode, resultCode, intent);
 
+        // Check if result comes from the correct activity
+        if (requestCode == REQUEST_CODE) {
+            AuthenticationResponse response = AuthenticationClient.getResponse(resultCode, intent);
 
-    // Create a request object with the optional parameter "market"
-    final GetSomethingRequest getSomethingRequest = spotifyApi.getSomething("qKRpDADUKrFeKhFHDMdfcu")
-            .market(CountryCode.SE)
-            .build();
+            switch (response.getType()) {
+                // Response was successful and contains auth token
+                case TOKEN:
+                    // Handle successful response
+                    break;
 
-    void getSomething_Sync() {
-        try {
-            // Execute the request synchronous
-            final Something something = getSomethingRequest.execute();
+                // Auth flow returned an error
+                case ERROR:
+                    // Handle error response
+                    break;
 
-            // Print something's name
-            System.out.println("Name: " + something.getName());
-        } catch (Exception e) {
-            System.out.println("Something went wrong!\n" + e.getMessage());
+                // Most likely auth flow was cancelled
+                default:
+                    // Handle other cases
+            }
         }
     }
 
-*/
+   @Override
+    protected void onStart() {
+        super.onStart();
+
+        SpotifyAppRemote.disconnect(mSpotifyAppRemote);
+
+        ConnectionParams connectionParams =
+                new ConnectionParams.Builder(CLIENT_ID)
+                        .setRedirectUri(REDIRECT_URI)
+                        .showAuthView(true)
+                        .build();
+
+        SpotifyAppRemote.connect(this, connectionParams,
+                new Connector.ConnectionListener() {
+
+                    @Override
+                    public void onConnected(SpotifyAppRemote spotifyAppRemote) {
+                        mSpotifyAppRemote = spotifyAppRemote;
+                        Log.d("MainActivity", "Connected! Yay!");
+
+                        // Now you can start interacting with App Remote
+
+                        connected();
+                    }
+
+                    @Override
+                    public void onFailure(Throwable throwable) {
+                        Log.e("MainActivity", throwable.getMessage(), throwable);
+
+                        // Something went wrong when attempting to connect! Handle errors here
+                    }
+                });
+
+
+   }
+
+    private void connected() {
+
+    }
+
+    @Override
+    protected void onStop() {
+        super.onStop();
+        SpotifyAppRemote.disconnect(mSpotifyAppRemote);
+    }
 }
