+++ QR/app/src/main/res/layout/activity_main.xml
@@ -92,16 +92,6 @@
         android:text="QueueName"
         android:textSize="22sp" />
 
-    <Button
-        android:id="@+id/btnTempMediaNav"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/rvUsers"
-        android:layout_alignParentStart="true"
-        android:layout_marginStart="99dp"
-        android:layout_marginTop="10dp"
-        android:text="Temporary Media Nav" />
-
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/bottomNav"
         android:layout_width="match_parent"
