+++ QR/app/src/main/res/layout/activity_media.xml
@@ -9,7 +9,7 @@
     <ImageView
         android:id="@+id/ivAlbum"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+        android:layout_height="143dp"
         android:layout_alignParentStart="true"
         android:layout_alignParentTop="true"
         android:layout_marginStart="18dp"
@@ -18,8 +18,8 @@
 
     <ImageView
         android:id="@+id/ivPlay"
-        android:layout_width="20dp"
-        android:layout_height="20dp"
+        android:layout_width="50dp"
+        android:layout_height="30dp"
         android:layout_below="@+id/sbSeeking"
         android:layout_marginStart="13dp"
         android:layout_marginTop="2dp"
@@ -28,8 +28,8 @@
 
     <ImageView
         android:id="@+id/ivPause"
-        android:layout_width="20dp"
-        android:layout_height="20dp"
+        android:layout_width="50dp"
+        android:layout_height="30dp"
         android:layout_below="@+id/sbSeeking"
         android:layout_marginStart="4dp"
         android:layout_marginTop="2dp"
@@ -45,7 +45,7 @@
         android:layout_marginTop="2dp"
         android:layout_toEndOf="@+id/ivAlbum"
         android:text="Album Name"
-        android:textSize="16sp" />
+        android:textSize="18sp" />
 
     <SeekBar
         android:id="@+id/sbSeeking"
@@ -65,7 +65,7 @@
         android:layout_marginTop="3dp"
         android:layout_toEndOf="@+id/ivAlbum"
         android:text="Artist Name"
-        android:textSize="18sp" />
+        android:textSize="20sp" />
 
     <TextView
         android:id="@+id/tvSongName"
@@ -76,7 +76,7 @@
         android:layout_marginTop="19dp"
         android:layout_toEndOf="@+id/ivAlbum"
         android:text="Song Title"
-        android:textSize="18sp" />
+        android:textSize="24sp" />
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvMusic"
